+++ mars_logging_android/CHANGELOG.md
@@ -1,3 +1,7 @@
+## 1.0.0-dev.1
+
+* Add compressMode, compressLevel, pubKey, maxFileSize & maxAliveDuration
+
 ## 1.0.0-dev.0
 
 * Add XLogImpl